Add new custom check for OfType
- OfType will ensure that the blocks in the project under test are in
the allowed list of resource Types specified in the `MatchSpec`
