+++ example/custom_ofType/main.tf
@@ -1,0 +1,7 @@
+resource "aws_sns_topic" "sns_topic" {
+    
+}
+
+resource "aws_instance" "some_instance" {
+    
+}
